Remove old source sets
